Refactor import statements in BaseClientModel.py and BaseServerModel.py
