+++ src/models/Server/BaseServerModel.py
@@ -8,7 +8,7 @@ from tools.CudaTool import get_device
 
 device = get_device()
 
-from models import BaseModel
+from models.BaseModel import BaseModel
 
 
 class BaseServerModel(BaseModel):
